Retain track parameters during add/remove track
Previously, actionAddTrackAbove()/Below() would always create the new
track at the very top of the layer stack (above all existing tracks),
then walk the list of those tracks' clips and slide them up one layer.

This caused _track_-level parameters (name, lock status) to be
left behind (in the case of lock status) or intentionally discarded
(in the case of track-name, which was blanked on all tracks which
were moved). Lock status was thus left behind on the "wrong track",
relative to the clips that track previously contained.

With this new code, the list of _tracks_ is walked during an add
operation, starting from the topmost track, and each track is moved
upwards one layer (with all of its properties intact). As each track
is moved, all of the clips on that track are pushed up along with it.
Then, a new track is created and slotted into the space vacated by
the lowermost moved track.

This method preserves all track parameters, because each track's data
in the project file is the _same_ after being moved as it was before.
Only the layer numbers change.

All of the label-blanking code was removed from both add methods,
as well as from the renumbering (in the other direction) performed
during actionRemoveTrack().