Avoid changing main window state during tutorial
The tutorial's event filter was using window.showNormal() to
activate the main OpenShot window, which forces the window into
a normal (as opposed to minimized/maximized/fullscreened) state.
This is disconcerting for users who wish to keep the application
in the state they've selected. Switch to window.show() which
respects existing state.

Fixes: #307, #1036, #1447, probably more duplicates