Use palette color for tutorial background
The tutorial windows were previously being painted with a hardcoded
background color of dark gray. When the theme was switched from
Humanity:Dark to Humanity, the background color stayed the same, but
the foreground text turned from white to black, making it difficult
to read.

This change uses the QPalette.Window color for the tutorial background,
which respects the current theme.

Humanity:Dark's tutorials will now be _slightly_ darker than the
previously-hardcoded background color, but not to a glaring degree
(though it is noticeable).