Make use of new translations resource
This modifies the translation-loading code to make use of the
":/locale" resource path for OpenShot translations, and greatly
simplifies the loading code by making use of Qt's automatic
fallback in `QTranslator.load()` instead of parsing the pieces
of the locale string directly.

The list of supported languages is now built using a QDirIterator
over the ":/locale" resource path.