Use track names in "Add To Timeline" window
I happened to notice that the track selector in the "Add To Timeline"
window (right-click a Project Files entry, select "Add To Timeline...')
didn't use the user-defined track labels, it always showed "Track _"
numbered tracks even when a label was set. There was no reason _not_
to use the labels, so now it does.