Add 256x256 PNG icon to xdg/ dir
Gnome just released 3.30, and with it some [new Design Guidelines][1]
to prepare for 3.32 (when they'll go into effect).

One of them is a new requirement that applications provide a 256x256
pixel icon for their application, to go along with the previous 48x48
icon. (OpenShot has been shipping a 64x64 icon, which is fine as it
gets scaled down when necessary.)

Since scaling the 64x64 icon _up_ to 256x256 looks **horrible**, this
change adds `xdg/openshot-qt_256.png` to the repo, exported from the
scalable SVG source just like the 64x64 icon.

[1]: https://help.gnome.org/misc/release-notes/3.30/developers.html.en#design-guidelines