Append correct extension to snapshot frames
Reading the code for `actionSaveFrame_trigger()`, I happened to notice 
what looks like a minor copy-paste bug. If the save filename chosen by 
the user doesn't have a `.png` extension, instead of adding it, the code 
would've accidentally been appending an `.osp` extension (OpenShot 
project file) instead.

Atom also caught a spurious space character following an `if ...:` 
statement, so I let it fix that as well. That's the other hunk in this 
PR.