preferences: Tweak item layout
The code for building the preferences window layout from the
available settings items was always trying to add `extraWidget`
to the item's `layout_hbox`, but since `extraWidget` is only
set on `param["type"] == "browse"` items, most of the time
`extraWidget` is undefined. This was causing a flurry of (harmless)
log messages whenever the prefs window was opened:
```
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
QLayout: Cannot add a null widget to QHBoxLayout/
```

This change wraps an `if(extraWidget):` around the code adding
it to the layout, so it will only be attempted when it exists.