Fix cursor and cursor-line shift
At far right position of the horizontal scroll the cursor and
cursor-line breaks apart and gets amall horizontal shift. The canvas of
the ruler rendered slightly bigger to eliminate this shift.

When width of the vertical scroll was enlarged - the corresponding
final width of the ruler needs to be enlarged to fully cover new width
of the scroll bars.