project.get(): Accept non-list args
previously project.get() required its argument to be a list, or it
would bail with a warning. This created lots of awkward code like:
`get_app().project.get(["layers"])`. Now, if a non-empty, non-list
arg is passed in, it simply promotes it to a list and continues,
making `get_app().project.get("layers")` valid and equivalent.