Connect resource file to consumers
I realized that, while it won't be an issue at runtime because we know 
the compiled resources will be loaded, it still makes sense to connect 
resources used in the ui files to the resource file itself. If nothing 
else, so that the resources are visible in Qt Creator / Designer.