Titles: Font selector reuses selection
Open the font selection dialog with the previous selection as the default, instead of always resetting to the default GUI font (Ubuntu).

When the window is created or a new file is loaded, we construct a default QFont() from the "Bitstream Vera Sans" typeface to use as the initial selection. Just as with the color picker, because we are not actully parsing the information from the data file, this initial selection may be incorrect... especially if the title file is a user-customized copy of a template.