Finished changes suggested by ferdnyc
The biggest change this time is that main_window.py now has a state variable indicating whether the timeline should follow the playhead when it is moved.  If it is set to True before the SeekSignal is emitted, the timeline will center on the playhead if it moves out of view.  The state variable is reset to False after the SeekSignal is handled.  This feels like a hack, but I can't figure out a better way to implement this functionality.

The timeline now centers on the playhead whenever it is actively moved by the user but not when it moves to follow the video player.