Files: Move add_file, get_image_seq() into model
Major overhaul to the code for adding items to the file model.
- `add_file()` is gone from both views
- files_model's new `add_files()` (<= note the plural) takes a list of
  paths, instead of churning over them one at a time
- Image sequence management is also consolidated into files_model
- Two helper methods in main_window handle communication with the user
  (in the form of popup questions / message boxes) for `add_files()`

Also, QPersistentModelIndex()es are stored in the custom model outer
classes' in-memory hashes, and can be used to locate a given item's
row for later modification (or deletion, specifically). No more
having to iterate over all the items to find one! (This is done for
Transitions as well, though it's currently not used there.)