About: Use BOM to decode UTF-16 changelog file (#3601)
If the Windows changelog text file (which is encoded UTF-16LE with
a Byte Order Marker at the start) is explicitly opened as 'utf_16_le',
the BOM will be left in the stream, which throws off the parsing of
the first line. Using 'encoding="utf_16"' automatically filters out
the BOM after using it to determine the file endianness.